Refactor edge expansion into extract phase.  New temporary file is generated - '.osrm.ebg' which is used by
