+++ contractor/contractor_options.cpp
@@ -48,9 +48,6 @@ ContractorOptions::ParseArguments(int argc, char *argv[], ContractorConfig &cont
     // declare a group of options that will be allowed both on command line and in config file
     boost::program_options::options_description config_options("Configuration");
     config_options.add_options()(
-        "restrictions,r",
-        boost::program_options::value<boost::filesystem::path>(&contractor_config.restrictions_path),
-        "Restrictions file in .osrm.restrictions format")(
         "profile,p", boost::program_options::value<boost::filesystem::path>(&contractor_config.profile_path)
                          ->default_value("profile.lua"),
         "Path to LUA routing profile")(
@@ -114,11 +111,6 @@ ContractorOptions::ParseArguments(int argc, char *argv[], ContractorConfig &cont
 
     boost::program_options::notify(option_variables);
 
-    if (!option_variables.count("restrictions"))
-    {
-        contractor_config.restrictions_path = contractor_config.osrm_input_path.string() + ".restrictions";
-    }
-
     if (!option_variables.count("input"))
     {
         SimpleLogger().Write() << "\n" << visible_options;
@@ -130,11 +122,7 @@ ContractorOptions::ParseArguments(int argc, char *argv[], ContractorConfig &cont
 
 void ContractorOptions::GenerateOutputFilesNames(ContractorConfig &contractor_config)
 {
-    contractor_config.node_output_path = contractor_config.osrm_input_path.string() + ".nodes";
     contractor_config.core_output_path = contractor_config.osrm_input_path.string() + ".core";
-    contractor_config.edge_output_path = contractor_config.osrm_input_path.string() + ".edges";
-    contractor_config.geometry_output_path = contractor_config.osrm_input_path.string() + ".geometry";
     contractor_config.graph_output_path = contractor_config.osrm_input_path.string() + ".hsgr";
-    contractor_config.rtree_nodes_output_path = contractor_config.osrm_input_path.string() + ".ramIndex";
-    contractor_config.rtree_leafs_output_path = contractor_config.osrm_input_path.string() + ".fileIndex";
+    contractor_config.edge_based_graph_path = contractor_config.osrm_input_path.string() + ".ebg";
 }
