+++ contractor/contractor_options.hpp
@@ -45,16 +45,11 @@ struct ContractorConfig
 
     boost::filesystem::path config_file_path;
     boost::filesystem::path osrm_input_path;
-    boost::filesystem::path restrictions_path;
     boost::filesystem::path profile_path;
 
-    std::string node_output_path;
     std::string core_output_path;
-    std::string edge_output_path;
-    std::string geometry_output_path;
     std::string graph_output_path;
-    std::string rtree_nodes_output_path;
-    std::string rtree_leafs_output_path;
+    std::string edge_based_graph_path;
 
     unsigned requested_num_threads;
