+++ test/.stxxl
@@ -1,1 +1,1 @@
-disk=/tmp/stxxl,10,syscall
+disk=###.stxxl,20,memory
