+++ util/fingerprint.hpp
@@ -52,7 +52,6 @@ class FingerPrint
 
     // initialize to {6ba7b810-9dad-11d1-80b4-00c04fd430c8}
     boost::uuids::uuid named_uuid;
-    bool has_64_bits;
 
 };
